+++ beacon-chain/light-update/event_debug.go
