+++ beacon-chain/light-update/event_test.go
