+++ beacon-chain/node/node.go
@@ -28,7 +28,6 @@ import (
 	"github.com/prysmaticlabs/prysm/beacon-chain/forkchoice"
 	"github.com/prysmaticlabs/prysm/beacon-chain/forkchoice/protoarray"
 	"github.com/prysmaticlabs/prysm/beacon-chain/gateway"
-	"github.com/prysmaticlabs/prysm/beacon-chain/light"
 	"github.com/prysmaticlabs/prysm/beacon-chain/node/registration"
 	"github.com/prysmaticlabs/prysm/beacon-chain/operations/attestations"
 	"github.com/prysmaticlabs/prysm/beacon-chain/operations/slashings"
@@ -194,11 +193,11 @@ func New(cliCtx *cli.Context, opts ...Option) (*BeaconNode, error) {
 		return nil, err
 	}
 
-	if err := beacon.registerLightClientService(); err != nil {
+	if err := beacon.registerSlasherService(); err != nil {
 		return nil, err
 	}
 
-	if err := beacon.registerSlasherService(); err != nil {
+	if err := beacon.registerLightUpdateService(); err != nil {
 		return nil, err
 	}
 
@@ -686,11 +685,6 @@ func (b *BeaconNode) registerRPCService() error {
 		return err
 	}
 
-	var lightService *light.Service
-	if err := b.services.FetchService(&lightService); err != nil {
-		return err
-	}
-
 	var slasherService *slasher.Service
 	if features.Get().EnableSlasher {
 		if err := b.services.FetchService(&slasherService); err != nil {
@@ -729,7 +723,7 @@ func (b *BeaconNode) registerRPCService() error {
 	}
 
 	p2pService := b.fetchP2P()
-	lightClientService := b.fetchLightClientService()
+	lightUpdateService := b.fetchLightUpdateService()
 	rpcService := rpc.NewService(b.ctx, &rpc.Config{
 		Host:                    host,
 		Port:                    port,
@@ -768,7 +762,7 @@ func (b *BeaconNode) registerRPCService() error {
 		StateGen:                b.stateGen,
 		EnableDebugRPCEndpoints: enableDebugRPCEndpoints,
 		MaxMsgSize:              maxMsgSize,
-		LightClientService:      lightClientService,
+		LightUpdateService:      lightUpdateService,
 	})
 
 	return b.services.RegisterService(rpcService)
@@ -877,35 +871,4 @@ func (b *BeaconNode) registerDeterminsticGenesisService() error {
 		return b.services.RegisterService(svc)
 	}
 	return nil
-}
-
-func (b *BeaconNode) registerLightClientService() error {
-	var chainService *blockchain.Service
-	if err := b.services.FetchService(&chainService); err != nil {
-		return err
-	}
-	var syncService *initialsync.Service
-	if err := b.services.FetchService(&syncService); err != nil {
-		return err
-	}
-	svc := light.New(b.ctx, &light.Config{
-		Database:                    b.db,
-		StateGen:                    b.stateGen,
-		HeadFetcher:                 chainService,
-		FinalizationFetcher:         chainService,
-		StateNotifier:               b,
-		TimeFetcher:                 chainService,
-		SyncChecker:                 syncService,
-		LightClientUpdatesQueueSize: 32, // TODO: jin; allowed this to be passed in from command line
-	})
-
-	return b.services.RegisterService(svc)
-}
-
-func (b *BeaconNode) fetchLightClientService() *light.Service {
-	var s *light.Service
-	if err := b.services.FetchService(&s); err != nil {
-		panic(err)
-	}
-	return s
 }
