+++ beacon-chain/rpc/prysm/v1alpha1/light-update/BUILD.bazel
