+++ beacon-chain/rpc/service.go
@@ -20,7 +20,7 @@ import (
 	opfeed "github.com/prysmaticlabs/prysm/beacon-chain/core/feed/operation"
 	statefeed "github.com/prysmaticlabs/prysm/beacon-chain/core/feed/state"
 	"github.com/prysmaticlabs/prysm/beacon-chain/db"
-	lightservice "github.com/prysmaticlabs/prysm/beacon-chain/light"
+	light_update "github.com/prysmaticlabs/prysm/beacon-chain/light-update"
 	"github.com/prysmaticlabs/prysm/beacon-chain/operations/attestations"
 	"github.com/prysmaticlabs/prysm/beacon-chain/operations/slashings"
 	"github.com/prysmaticlabs/prysm/beacon-chain/operations/synccommittee"
@@ -34,7 +34,7 @@ import (
 	"github.com/prysmaticlabs/prysm/beacon-chain/rpc/eth/validator"
 	beaconv1alpha1 "github.com/prysmaticlabs/prysm/beacon-chain/rpc/prysm/v1alpha1/beacon"
 	debugv1alpha1 "github.com/prysmaticlabs/prysm/beacon-chain/rpc/prysm/v1alpha1/debug"
-	"github.com/prysmaticlabs/prysm/beacon-chain/rpc/prysm/v1alpha1/light"
+	light_update_server "github.com/prysmaticlabs/prysm/beacon-chain/rpc/prysm/v1alpha1/light-update"
 	nodev1alpha1 "github.com/prysmaticlabs/prysm/beacon-chain/rpc/prysm/v1alpha1/node"
 	validatorv1alpha1 "github.com/prysmaticlabs/prysm/beacon-chain/rpc/prysm/v1alpha1/validator"
 	"github.com/prysmaticlabs/prysm/beacon-chain/rpc/statefetcher"
@@ -110,7 +110,7 @@ type Config struct {
 	OperationNotifier       opfeed.Notifier
 	StateGen                *stategen.State
 	MaxMsgSize              int
-	LightClientService      *lightservice.Service
+	LightUpdateService      *light_update.Service
 }
 
 // NewService instantiates a new RPC service instance that will
@@ -283,11 +283,11 @@ func (s *Service) Start() {
 		V1Alpha1ValidatorServer: validatorServer,
 	}
 
-	lightClientServer := &light.Server{
-		LightClientService: s.cfg.LightClientService,
+	lightUpdateServer := &light_update_server.Server{
+		LightUpdateService: s.cfg.LightUpdateService,
 	}
 
-	ethpbv1alpha1.RegisterLightClientServer(s.grpcServer, lightClientServer)
+	ethpbv1alpha1.RegisterLightClientServer(s.grpcServer, lightUpdateServer)
 	ethpbv1alpha1.RegisterNodeServer(s.grpcServer, nodeServer)
 	ethpbservice.RegisterBeaconNodeServer(s.grpcServer, nodeServerV1)
 	ethpbv1alpha1.RegisterHealthServer(s.grpcServer, nodeServer)
