+++ go.mod
@@ -16,11 +16,6 @@ require (
 	github.com/dgraph-io/ristretto v0.0.4-0.20210318174700-74754f61e018
 	github.com/dustin/go-humanize v1.0.0
 	github.com/emicklei/dot v0.11.0
-	// TODO:
-	// - not sure if this is a good choice. We just need a simple FIFO queue;
-	// - do we need this to be concurrent?
-	// - Use a simple, first principle implementation + a mutex write for a single writer pattern
-	github.com/enriquebris/goconcurrentqueue v0.6.0 // indirect
 	github.com/ethereum/go-ethereum v1.10.10
 	github.com/fatih/color v1.9.0 // indirect
 	github.com/ferranbt/fastssz v0.0.0-20210905181407-59cf6761a7d5
